django/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.28.81
botocore==1.31.81
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.7
django-allow-cidr==0.7.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-filter==23.3
django-graphql-jwt==0.4.0
django-storages==1.14.2
django-timezone-field==6.0.1
dnspython==2.4.2
email-validator==2.0.0
exceptiongroup==1.1.3
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.3
packaging==23.2
pluggy==1.3.0
promise==2.3
prompt-toolkit==3.0.39
psycopg==3.1.12
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
redis==4.6.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.9
